Move code to social_users, rewrite FVB to CBV
