+++ django_project/social_users/urls.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from django.conf.urls import include, patterns, url
 
-from .views import LogoutUser, ProfilePage, UserProfilePage, UserSigninPage
+from .views import GetUserUpdates, LogoutUser, ProfilePage, UserProfilePage, UserSigninPage
 
 urlpatterns = patterns(
     '',
@@ -14,8 +14,6 @@ urlpatterns = patterns(
     ),
     url(r'^logout/$', LogoutUser.as_view(), name='logout_user'),
     url(
-        r'^user/updates/$',
-        'social_users.views.get_user_updates',
-        name='user-updates'
+        r'^user/updates/$', GetUserUpdates.as_view(), name='user-updates'
     ),
 )
